+++ packages/tsurlfilter/src/rules/cosmetic-rule-parser.ts
@@ -140,7 +140,7 @@ export class CosmeticRuleParser {
             const modifierText = modifiersTextArray[i];
             const assignerIndex = modifierText.indexOf(Assigner);
 
-            if (modifierText === 'path') {
+            if (modifierText === CosmeticRuleModifiers.Path) {
                 // Empty path modifier without assigner and value will match only main page
                 modifiers[modifierText] = '';
                 continue;
@@ -151,6 +151,10 @@ export class CosmeticRuleParser {
             }
 
             const modifierKey = modifierText.substring(0, assignerIndex);
+
+            if (modifierKey === CosmeticRuleModifiers.Url && modifiersTextArray.length > 1) {
+                throw new SyntaxError('The $url modifier can\'t be used with other modifiers');
+            }
 
             if (cosmeticRuleModifiersList.includes(modifierKey)) {
                 const modifierValue = modifierText.substring(assignerIndex + 1);
@@ -169,10 +173,12 @@ export class CosmeticRuleParser {
      * If domains are declared through $domain modifier and pattern domain list, this method throws a SyntaxError.
      * @param rulePattern - rule pattern text
      *
-     * @returns Object with permitted/restricted domains list and the path modifier string value
+     * @returns Object with permitted/restricted domains list and/or the path modifier string value,
+     * or url modifier string value
      */
     static parseRulePattern(rulePattern: string): {
-        path?: string;
+        url?: string,
+        path?: string,
         permittedDomains?: string[];
         restrictedDomains?: string[];
     } {
@@ -187,6 +193,15 @@ export class CosmeticRuleParser {
         const modifiers = CosmeticRuleParser.parseRuleModifiers(modifiersText);
 
         if (modifiers) {
+            if (modifiers.url) {
+                if (domains) {
+                    throw new SyntaxError('The $url modifier is not allowed in a domain-specific rule');
+                } else {
+                    const { url } = modifiers;
+                    return { url };
+                }
+            }
+
             if (modifiers.path || modifiers.path === '') {
                 path = modifiers.path;