+++ packages/tsurlfilter/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 - New `$permissions` modifier to set Permissions-Policy response header.
 - Support conversion to DNR for `$cookie` modifier.
+- `$url` modifier support for non-basic rules
 
 ## [2.1.5] - 2023-07-21